+++ policyengine_us/variables/household/healthcare_benefit_value.py
@@ -21,6 +21,8 @@ class healthcare_benefit_value(Variable):
         "assigned_ca_premium_subsidy",
         "basic_health_program",
         "co_omnisalud",
+        "assigned_co_premium_assistance",
         "md_premium_assistance",
+        "assigned_nm_premium_assistance",
         "or_healthier_oregon_cost",
     ]